+++ cr8/bench_spec.py
@@ -1,12 +1,8 @@
 
 import os
+from os.path import basename
 import json
 import toml
-import pathlib
-
-
-def file_name(full_path):
-    return pathlib.Path(full_path).name
 
 
 class Instructions:
@@ -28,32 +24,35 @@ class Instructions:
 
 
 class Spec:
-    def __init__(self, setup, teardown, name=None, queries=None, load_data=None):
+    def __init__(self, setup, teardown, queries=None, load_data=None, meta=None, **metadata):
         self.setup = setup
         self.teardown = teardown
         self.queries = queries
         self.load_data = load_data
-        self.name = name
+        self.meta = metadata
+        if meta:
+            self.meta.update(meta)
 
     @staticmethod
-    def from_dict(d, name=None):
+    def from_dict(d, **metadata):
         return Spec(
             setup=Instructions.from_dict(d.get('setup', {})),
             teardown=Instructions.from_dict(d.get('teardown', {})),
-            name=d.get('name', name),
+            meta=d.get('meta', {}),
             queries=d.get('queries', []),
-            load_data=d.get('load_data', [])
+            load_data=d.get('load_data', []),
+            **metadata
         )
 
     @staticmethod
     def from_json_file(filename):
         with open(filename, 'r', encoding='utf-8') as spec_file:
-            return Spec.from_dict(json.load(spec_file), name=file_name(filename))
+            return Spec.from_dict(json.load(spec_file), name=basename(filename))
 
     @staticmethod
     def from_toml_file(filename):
         with open(filename, 'r', encoding='utf-8') as spec_file:
-            return Spec.from_dict(toml.loads(spec_file.read()), name=file_name(filename))
+            return Spec.from_dict(toml.loads(spec_file.read()), name=basename(filename))
 
     @staticmethod
     def from_python_file(filename):
@@ -65,8 +64,7 @@ class Spec:
             code = compile(f.read(), filename, 'exec')
             exec(code, global_vars)
             spec = global_vars['spec']
-            if not spec.name:
-                spec.name = file_name(filename)
+            spec.meta['name'] = spec.meta.get('name', basename(filename))
             return spec