+++ cr8/run_spec.py
@@ -22,7 +22,9 @@ create table if not exists benchmarks (
         hash string
     ),
     statement string,
-    spec_name string,
+    meta object as (
+        name string
+    ),
     started timestamp,
     ended timestamp,
     concurrency int,
@@ -107,7 +109,7 @@ class Executor:
             aio.run_many(self.client.execute_many, inserts, concurrency=concurrency)
             cursor.execute('refresh table {target}'.format(target=data_file['target']))
 
-    def run_load_data(self, data_spec, spec_name=None):
+    def run_load_data(self, data_spec, meta=None):
         inserts = self._to_inserts(data_spec)
         statement = next(iter(inserts))[0]
         bulk_size = data_spec.get('bulk_size', 5000)
@@ -127,7 +129,7 @@ class Executor:
         self.process_result(Result(
             version_info=self.server_version_info,
             statement=statement,
-            spec_name=spec_name,
+            meta=meta,
             started=start,
             ended=end,
             stats=stats,
@@ -145,7 +147,7 @@ class Executor:
             server_version='.'.join((str(x) for x in self.server_version)))
         return msg
 
-    def run_queries(self, queries, spec_name=None):
+    def run_queries(self, queries, meta=None):
         for query in queries:
             stmt = query['statement']
             iterations = query.get('iterations', 1)
@@ -163,7 +165,7 @@ class Executor:
                        concurrency=concurrency)))
             with QueryRunner(
                 stmt,
-                spec_name,
+                meta,
                 repeats=iterations,
                 hosts=self.benchmark_hosts,
                 concurrency=concurrency,
@@ -227,9 +229,9 @@ def run_spec(spec, benchmark_hosts, result_hosts=None, output_fmt=None):
             print('# Running benchmark')
             if spec.load_data:
                 for data_spec in spec.load_data:
-                    executor.run_load_data(data_spec, spec.name)
+                    executor.run_load_data(data_spec, spec.meta)
             else:
-                executor.run_queries(spec.queries, spec.name)
+                executor.run_queries(spec.queries, spec.meta)
         finally:
             print('# Running tearDown')
             executor.exec_instructions(spec.teardown)